Batched delivery.
This implements full batched delivery.  Batches are sent depending on
several tunables - a time based value, a count based one, and a size based
one.  There is also explicit flushing.  Even single events will be sent
using the batch mechanism for now.  Default thresholds are 10 seconds, 250 events,
or 500 KB.

There is no support for adding Context or Integrations objects to the batch
at this time.  Doing that will require exposing a new API that allows consumers
to attach that to the Analytics object.