Significant API improvements.
This includes a vanilla Object API for passing around Context and
Integrations objects, and the ability to access all the relevant
fields for every API call.  It also arranges to populate the timestamp
and sentAt fields correctly (accounting for queueing) and offers a
convenience method for users to obtain the current time as a nicely
formatted ISO8601 string.