Substantial fixes for Windows.
When we switched to start testing with AppVeyor, I found many problems.
There appears to be a subtle race condition somewhere in the Windows
stack, but adding a brief delay (10ms) clears it and keeps the queue
from getting stuck.

While here, we have added a native WinInet transport which is used by
default on Windows instead of CURL, and CMake support to select it,
as well as a switch to remove all HTTP transports (-DNO_DEFAULT_HTTP=ON).